Moved handleUserData to src/utils folder to reuse it
