+++ src/pages/User/index.vue
@@ -3,6 +3,10 @@
 </template>
 
 <script setup lang="ts">
+import { User } from 'src/utils';
+import { ref } from 'vue';
 import UserLayout from './User.layout.vue';
-import { user } from './mock';
+import { getUser } from 'src/utils';
+
+const user = ref<User>(getUser());
 </script>
