+++ src/utils/index.js
@@ -5,3 +5,4 @@ export * from './validations';
 export * from './shoppingCart';
 export * from './confirmBeforeExit';
 export * from './auth';
+export * from './handleUserData';
